+++ spec/rspec/core/example_group_spec.rb
@@ -292,7 +292,7 @@ module RSpec::Core
         context "and a string in an inner group" do
           it "is the top level constant" do
             group = ExampleGroup.describe(String) do
-              describe :symbol do
+              describe "some string" do
                 example "described_class is String" do
                   expect(described_class).to eq(String)
                 end
@@ -308,7 +308,7 @@ module RSpec::Core
             group = ExampleGroup.describe(String) do
               described_class
               metadata[:example_group][:described_class] = Object
-              describe :symbol do
+              describe "inner" do
                 example "described_class is Object" do
                   expect(described_class).to eq(Object)
                 end